refactor & update write interface
