+++ experiments/lmdb-oxen/src/explore.rs
@@ -1,5 +1,5 @@
 pub mod interfaces;
-pub mod read_lazy_merkle;
+pub mod merkle_reader;
 pub mod lazy_merkle_lmdb_stateful;
 pub mod lmdb_impl;
 pub mod merkle_writer;
@@ -7,3 +7,4 @@ pub mod migration_explore;
 pub mod new_main;
 pub mod new_path;
 pub mod scratch;
+pub mod lazy_merkle;
